Move search query into newly created search app
